+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Changes to AdfsOrganization
   - Added empty organization support ([issue #30](https://github.com/X-Guardian/AdfsDsc/issues/30)).
   - Added integration tests.
+- Changes to AdfsProperties
+  - Added integration tests.
 
 ## 1.0.0